- рефакторинг - PatientRecordServiceTest - IcdServiceTest - ExaminationServiceTest - DeviceServiceTest - ChannelNameServiceTest

+++ src/main/java/ru/gsa/biointerface/services/ChannelNameService.java
@@ -4,10 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.gsa.biointerface.domain.entity.ChannelName;
 import ru.gsa.biointerface.repository.ChannelNameRepository;
-import ru.gsa.biointerface.repository.exception.NoConnectionException;
+import ru.gsa.biointerface.repository.exception.InsertException;
 
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.NonUniqueResultException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -19,16 +18,16 @@ public class ChannelNameService {
     private static ChannelNameService instance = null;
     private final ChannelNameRepository dao;
 
-    public static ChannelNameService getInstance() throws NoConnectionException {
+    private ChannelNameService() throws Exception {
+        dao = ChannelNameRepository.getInstance();
+    }
+
+    public static ChannelNameService getInstance() throws Exception {
         if (instance == null) {
             instance = new ChannelNameService();
         }
 
         return instance;
-    }
-
-    private ChannelNameService() throws NoConnectionException {
-        dao = ChannelNameRepository.getInstance();
     }
 
     public ChannelName create(String name, String comment) throws Exception {
@@ -43,10 +42,10 @@ public class ChannelNameService {
         return entity;
     }
 
-    public List<ChannelName> getAll() throws Exception{
+    public List<ChannelName> getAll() throws Exception {
         List<ChannelName> entities = dao.getAll();
 
-        if(entities.size() > 0) {
+        if (entities.size() > 0) {
             LOGGER.info("Get all channelNames from database");
         } else {
             LOGGER.info("ChannelNames is not found in database");
@@ -56,9 +55,12 @@ public class ChannelNameService {
     }
 
     public ChannelName getById(long id) throws Exception {
+        if (id <= 0)
+            throw new IllegalArgumentException("id <= 0");
+
         ChannelName entity = dao.read(id);
 
-        if(entity != null) {
+        if (entity != null) {
             LOGGER.info("Get channelName(id={}) from database", entity.getId());
         } else {
             LOGGER.error("ChannelName(id={}) is not found in database", id);
@@ -69,8 +71,12 @@ public class ChannelNameService {
     }
 
     public void save(ChannelName entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("entity is null");
         if (entity.getName() == null)
             throw new NullPointerException("Name is null");
+        if (entity.getChannels() == null)
+            throw new NullPointerException("Channels is null");
         if ("".equals(entity.getName()))
             throw new IllegalArgumentException("Name is empty");
 
@@ -81,14 +87,19 @@ public class ChannelNameService {
             LOGGER.info("ChannelName(id={})  is recorded in database", entity.getId());
         } else {
             LOGGER.error("ChannelName(id={}) already exists in database", entity.getId());
-            throw new NonUniqueResultException("ChannelName(id=" + entity.getId() + ") already exists in database");
+            throw new InsertException("ChannelName(id=" + entity.getId() + ") already exists in database");
         }
     }
 
     public void delete(ChannelName entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("entity is null");
+        if (entity.getId() <= 0)
+            throw new IllegalArgumentException("id <= 0");
+
         ChannelName readEntity = dao.read(entity.getId());
 
-        if(readEntity != null) {
+        if (readEntity != null) {
             dao.delete(entity);
             LOGGER.info("ChannelName(id={}) is deleted in database", entity.getId());
         } else {
@@ -98,14 +109,20 @@ public class ChannelNameService {
     }
 
     public void update(ChannelName entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("entity is null");
         if (entity.getName() == null)
             throw new NullPointerException("Name is null");
+        if (entity.getChannels() == null)
+            throw new NullPointerException("Channels is null");
         if ("".equals(entity.getName()))
             throw new IllegalArgumentException("Name is empty");
+        if (entity.getId() <= 0)
+            throw new IllegalArgumentException("id <= 0");
 
         ChannelName readEntity = dao.read(entity.getId());
 
-        if(readEntity != null) {
+        if (readEntity != null) {
             dao.update(entity);
             LOGGER.info("ChannelName(id={}) updated in database", entity.getId());
         } else {
