+++ src/main/java/ru/gsa/biointerface/services/DeviceService.java
@@ -4,9 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.gsa.biointerface.domain.entity.Device;
 import ru.gsa.biointerface.repository.DeviceRepository;
-import ru.gsa.biointerface.repository.exception.NoConnectionException;
 
 import javax.persistence.EntityNotFoundException;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -17,7 +17,11 @@ public class DeviceService {
     private static DeviceService instance = null;
     private final DeviceRepository dao;
 
-    public static DeviceService getInstance() throws NoConnectionException {
+    private DeviceService() throws Exception {
+        dao = DeviceRepository.getInstance();
+    }
+
+    public static DeviceService getInstance() throws Exception {
         if (instance == null) {
             instance = new DeviceService();
         }
@@ -25,17 +29,13 @@ public class DeviceService {
         return instance;
     }
 
-    private DeviceService() throws NoConnectionException {
-        dao = DeviceRepository.getInstance();
-    }
+    public Device create(long id, int amountChannels) {
+        if (id <= 0)
+            throw new IllegalArgumentException("Serial number <= 0");
+        if (amountChannels <= 0)
+            throw new IllegalArgumentException("Amount channels <= 0");
 
-    public Device create(int id, int amountChannels) {
-        if (id < 0)
-            throw new IllegalArgumentException("Serial number is '0'");
-        if (amountChannels == 0)
-            throw new IllegalArgumentException("Amount channels is '0'");
-
-        Device entity = new Device(id, amountChannels, "");
+        Device entity = new Device(id, amountChannels, "", new ArrayList<>());
         LOGGER.info("New device created");
 
         return entity;
@@ -44,7 +44,7 @@ public class DeviceService {
     public List<Device> getAll() throws Exception {
         List<Device> entities = dao.getAll();
 
-        if(entities.size() > 0) {
+        if (entities.size() > 0) {
             LOGGER.info("Get all devices from database");
         } else {
             LOGGER.info("Devices is not found in database");
@@ -54,9 +54,12 @@ public class DeviceService {
     }
 
     public Device getById(long id) throws Exception {
+        if (id <= 0)
+            throw new IllegalArgumentException("id <= 0");
+
         Device entity = dao.read(id);
 
-        if(entity != null) {
+        if (entity != null) {
             LOGGER.info("Get device(id={}) from database", entity.getId());
         } else {
             LOGGER.error("Device(id={}) is not found in database", id);
@@ -67,10 +70,14 @@ public class DeviceService {
     }
 
     public void save(Device entity) throws Exception {
-        if (entity.getId() < 0)
-            throw new IllegalArgumentException("Serial number is '0'");
-        if (entity.getAmountChannels() == 0)
-            throw new IllegalArgumentException("Amount channels is '0'");
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
+        if (entity.getId() <= 0)
+            throw new IllegalArgumentException("id <= 0");
+        if (entity.getAmountChannels() <= 0)
+            throw new IllegalArgumentException("Amount channels <= 0");
+        if (entity.getExaminations() == null)
+            throw new NullPointerException("Examinations is null");
 
         Device readEntity = dao.read(entity.getId());
 
@@ -79,7 +86,7 @@ public class DeviceService {
             LOGGER.info("Device(id={}) is recorded in database", entity.getId());
         } else {
             LOGGER.warn("Device(id={}) already exists in database", entity.getId());
-            if(readEntity.getAmountChannels() != entity.getAmountChannels()){
+            if (readEntity.getAmountChannels() != entity.getAmountChannels()) {
                 LOGGER.error("Amount of channels does not match previously recorded");
                 throw new IllegalArgumentException("Amount of channels does not match previously recorded");
             }
@@ -87,9 +94,14 @@ public class DeviceService {
     }
 
     public void delete(Device entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
+        if (entity.getId() <= 0)
+            throw new IllegalArgumentException("id <= 0");
+
         Device readEntity = dao.read(entity.getId());
 
-        if(readEntity != null) {
+        if (readEntity != null) {
             dao.delete(entity);
             LOGGER.info("Device(id={}) is deleted in database", entity.getId());
         } else {
@@ -99,14 +111,18 @@ public class DeviceService {
     }
 
     public void update(Device entity) throws Exception {
-        if (entity.getId() < 0)
-            throw new IllegalArgumentException("Serial number is '0'");
-        if (entity.getAmountChannels() == 0)
-            throw new IllegalArgumentException("Amount channels is '0'");
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
+        if (entity.getId() <= 0)
+            throw new IllegalArgumentException("Id <= 0");
+        if (entity.getAmountChannels() <= 0)
+            throw new IllegalArgumentException("Amount channels <= 0");
+        if (entity.getExaminations() == null)
+            throw new NullPointerException("Examinations is null");
 
         Device readEntity = dao.read(entity.getId());
 
-        if(readEntity != null) {
+        if (readEntity != null) {
             dao.update(entity);
             LOGGER.info("Device(id={}) updated in database", entity.getId());
         } else {
