+++ src/main/java/ru/gsa/biointerface/services/IcdService.java
@@ -4,8 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.gsa.biointerface.domain.entity.Icd;
 import ru.gsa.biointerface.repository.IcdRepository;
-import ru.gsa.biointerface.repository.exception.NoConnectionException;
 
+import javax.persistence.EntityNotFoundException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -17,7 +18,11 @@ public class IcdService {
     private static IcdService instance = null;
     private final IcdRepository dao;
 
-    public static IcdService getInstance() throws NoConnectionException {
+    private IcdService() throws Exception {
+        dao = IcdRepository.getInstance();
+    }
+
+    public static IcdService getInstance() throws Exception {
         if (instance == null) {
             instance = new IcdService();
         }
@@ -25,28 +30,24 @@ public class IcdService {
         return instance;
     }
 
-    private IcdService() throws NoConnectionException {
-        dao = IcdRepository.getInstance();
-    }
-
     public Icd create(String name, int version, String comment) throws Exception {
         if (name == null)
             throw new NullPointerException("name is null");
         if ("".equals(name))
-            throw new NullPointerException("name is empty");
+            throw new IllegalArgumentException("name is empty");
         if (version <= 0)
             throw new IllegalArgumentException("version <= 0");
 
-        Icd entity = new Icd(-1, name, version, comment);
+        Icd entity = new Icd(-1, name, version, comment, new ArrayList<>());
         LOGGER.info("New icd created");
 
         return entity;
     }
 
-    public List<Icd> getAll()  throws Exception {
+    public List<Icd> getAll() throws Exception {
         List<Icd> entities = dao.getAll();
 
-        if(entities.size() > 0) {
+        if (entities.size() > 0) {
             LOGGER.info("Get all icds from database");
         } else {
             LOGGER.info("Icds is not found in database");
@@ -55,10 +56,13 @@ public class IcdService {
         return entities;
     }
 
-    public Icd getById(long id)  throws Exception {
+    public Icd getById(long id) throws Exception {
+        if (id <= 0)
+            throw new IllegalArgumentException("id <= 0");
+
         Icd entity = dao.read(id);
 
-        if(entity != null) {
+        if (entity != null) {
             LOGGER.info("Get icd(id={}) from database", entity.getId());
         } else {
             LOGGER.error("Icd(id={}) is not found in database", id);
@@ -68,13 +72,17 @@ public class IcdService {
         return entity;
     }
 
-    public void save(Icd entity)  throws Exception {
+    public void save(Icd entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
         if (entity.getName() == null)
             throw new NullPointerException("name is null");
         if ("".equals(entity.getName()))
-            throw new NullPointerException("name is empty");
+            throw new IllegalArgumentException("name is empty");
         if (entity.getVersion() <= 0)
             throw new IllegalArgumentException("version <= 0");
+        if (entity.getPatientRecords() == null)
+            throw new NullPointerException("PatientRecords is null");
 
         Icd readEntity = dao.read(entity.getId());
 
@@ -87,34 +95,45 @@ public class IcdService {
         }
     }
 
-    public void delete(Icd entity)  throws Exception {
+    public void delete(Icd entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
+        if (entity.getId() <= 0)
+            throw new IllegalArgumentException("Id <= 0");
+
         Icd readEntity = dao.read(entity.getId());
 
-        if(readEntity != null) {
+        if (readEntity != null) {
             dao.delete(entity);
             LOGGER.info("Icd(id={}) is deleted in database", entity.getId());
         } else {
             LOGGER.error("Icd(id={}) not found in database", entity.getId());
-            throw new NoSuchElementException("Icd(id=" + entity.getId() + ") not found in database");
+            throw new EntityNotFoundException("Icd(id=" + entity.getId() + ") not found in database");
         }
     }
 
-    public void update(Icd entity)  throws Exception {
+    public void update(Icd entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
+        if (entity.getId() <= 0)
+            throw new IllegalArgumentException("id <= 0");
         if (entity.getName() == null)
-            throw new NullPointerException("name is null");
+            throw new NullPointerException("Name is null");
         if ("".equals(entity.getName()))
-            throw new NullPointerException("name is empty");
+            throw new IllegalArgumentException("Name is empty");
         if (entity.getVersion() <= 0)
             throw new IllegalArgumentException("version <= 0");
+        if (entity.getPatientRecords() == null)
+            throw new NullPointerException("PatientRecords is null");
 
         Icd readEntity = dao.read(entity.getId());
 
-        if(readEntity != null) {
+        if (readEntity != null) {
             dao.update(entity);
             LOGGER.info("Icd(id={}) updated in database", entity.getId());
         } else {
             LOGGER.error("Icd(id={}) not found in database", entity.getId());
-            throw new NoSuchElementException("Icd(id=" + entity.getId() + ") not found in database");
+            throw new EntityNotFoundException("Icd(id=" + entity.getId() + ") not found in database");
         }
     }
 }
