+++ src/main/java/ru/gsa/biointerface/services/PatientRecordService.java
@@ -5,9 +5,10 @@ import org.slf4j.LoggerFactory;
 import ru.gsa.biointerface.domain.entity.Icd;
 import ru.gsa.biointerface.domain.entity.PatientRecord;
 import ru.gsa.biointerface.repository.PatientRecordRepository;
-import ru.gsa.biointerface.repository.exception.*;
 
+import javax.persistence.EntityNotFoundException;
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -20,56 +21,16 @@ public class PatientRecordService {
     private static PatientRecordService instance = null;
     private final PatientRecordRepository dao;
 
-    public static PatientRecordService getInstance() throws NoConnectionException {
+    private PatientRecordService() throws Exception {
+        dao = PatientRecordRepository.getInstance();
+    }
+
+    public static PatientRecordService getInstance() throws Exception {
         if (instance == null) {
             instance = new PatientRecordService();
         }
 
         return instance;
-    }
-
-    private PatientRecordService() throws NoConnectionException {
-        dao = PatientRecordRepository.getInstance();
-    }
-
-    public PatientRecord create(
-            int id,
-            String secondName,
-            String firstName,
-            String middleName,
-            LocalDate birthday,
-            Icd icd,
-            String comment
-    )  throws Exception {
-        if (id <= 0)
-            throw new NullPointerException("Id <= 0");
-        if (secondName == null)
-            throw new NullPointerException("SecondName is null");
-        if ("".equals(secondName))
-            throw new NullPointerException("SecondName is empty");
-        if (firstName == null)
-            throw new NullPointerException("FirstName is null");
-        if ("".equals(firstName))
-            throw new NullPointerException("FirstName is empty");
-        if (middleName == null)
-            throw new NullPointerException("MiddleName is null");
-        if ("".equals(middleName))
-            throw new NullPointerException("MiddleName is empty");
-        if (birthday == null)
-            throw new NullPointerException("Birthday is null");
-
-        PatientRecord entity = new PatientRecord(
-                id,
-                secondName,
-                firstName,
-                middleName,
-                localDateToDate(birthday),
-                icd,
-                comment
-        );
-        LOGGER.info("New patient record created");
-
-        return entity;
     }
 
     private static Calendar localDateToDate(LocalDate localDate) {
@@ -85,10 +46,50 @@ public class PatientRecordService {
         return calendar;
     }
 
+    public PatientRecord create(
+            long id,
+            String secondName,
+            String firstName,
+            String middleName,
+            LocalDate birthday,
+            Icd icd,
+            String comment
+    ) throws Exception {
+        if (id <= 0)
+            throw new IllegalArgumentException("Id <= 0");
+        if (secondName == null)
+            throw new NullPointerException("SecondName is null");
+        if ("".equals(secondName))
+            throw new IllegalArgumentException("SecondName is empty");
+        if (firstName == null)
+            throw new NullPointerException("FirstName is null");
+        if ("".equals(firstName))
+            throw new IllegalArgumentException("FirstName is empty");
+        if (middleName == null)
+            throw new NullPointerException("MiddleName is null");
+        if ("".equals(middleName))
+            throw new IllegalArgumentException("MiddleName is empty");
+        if (birthday == null)
+            throw new NullPointerException("Birthday is null");
+
+        PatientRecord entity = new PatientRecord(
+                id,
+                secondName,
+                firstName,
+                middleName,
+                localDateToDate(birthday),
+                icd,
+                comment,
+                new ArrayList<>());
+        LOGGER.info("New patient record created");
+
+        return entity;
+    }
+
     public List<PatientRecord> getAll() throws Exception {
         List<PatientRecord> entities = dao.getAll();
 
-        if(entities.size() > 0) {
+        if (entities.size() > 0) {
             LOGGER.info("Get all patientRecords from database");
         } else {
             LOGGER.info("PatientRecords is not found in database");
@@ -98,13 +99,16 @@ public class PatientRecordService {
     }
 
     public PatientRecord getById(long id) throws Exception {
+        if (id <= 0)
+            throw new IllegalArgumentException("id <= 0");
+
         PatientRecord entity = dao.read(id);
 
-        if(entity != null) {
+        if (entity != null) {
             LOGGER.info("Get patientRecord(id={}) from database", entity.getId());
         } else {
             LOGGER.error("PatientRecord(id={}) is not found in database", id);
-            throw new NoSuchElementException(
+            throw new EntityNotFoundException(
                     "PatientRecord(id=" + id + ") is not found in database"
             );
         }
@@ -113,22 +117,26 @@ public class PatientRecordService {
     }
 
     public void save(PatientRecord entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
         if (entity.getId() <= 0)
-            throw new NullPointerException("Id <= 0");
+            throw new IllegalArgumentException("Id <= 0");
         if (entity.getSecondName() == null)
             throw new NullPointerException("SecondName is null");
         if ("".equals(entity.getSecondName()))
-            throw new NullPointerException("SecondName is empty");
+            throw new IllegalArgumentException("SecondName is empty");
         if (entity.getFirstName() == null)
             throw new NullPointerException("FirstName is null");
         if ("".equals(entity.getFirstName()))
-            throw new NullPointerException("FirstName is empty");
+            throw new IllegalArgumentException("FirstName is empty");
         if (entity.getMiddleName() == null)
             throw new NullPointerException("MiddleName is null");
         if ("".equals(entity.getMiddleName()))
-            throw new NullPointerException("MiddleName is empty");
+            throw new IllegalArgumentException("MiddleName is empty");
         if (entity.getBirthday() == null)
             throw new NullPointerException("Birthday is null");
+        if (entity.getExaminations() == null)
+            throw new NullPointerException("Examinations is null");
 
         PatientRecord readEntity = dao.read(entity.getId());
 
@@ -143,41 +151,50 @@ public class PatientRecordService {
         }
     }
 
-    public void delete(PatientRecord entity)  throws Exception {
+    public void delete(PatientRecord entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
+        if (entity.getId() <= 0)
+            throw new IllegalArgumentException("id <= 0");
+
         PatientRecord readEntity = dao.read(entity.getId());
 
-        if(readEntity != null) {
+        if (readEntity != null) {
             dao.delete(entity);
             LOGGER.info("PatientRecord(id={}) is deleted in database", entity.getId());
         } else {
             LOGGER.error("PatientRecord(id={}) not found in database", entity.getId());
-            throw new NoSuchElementException(
+            throw new EntityNotFoundException(
                     "PatientRecord(id=" + entity.getId() + ") not found in database"
             );
         }
     }
 
     public void update(PatientRecord entity) throws Exception {
+        if (entity == null)
+            throw new NullPointerException("Entity is null");
         if (entity.getId() <= 0)
-            throw new NullPointerException("Id <= 0");
+            throw new IllegalArgumentException("Id <= 0");
         if (entity.getSecondName() == null)
             throw new NullPointerException("SecondName is null");
         if ("".equals(entity.getSecondName()))
-            throw new NullPointerException("SecondName is empty");
+            throw new IllegalArgumentException("SecondName is empty");
         if (entity.getFirstName() == null)
             throw new NullPointerException("FirstName is null");
         if ("".equals(entity.getFirstName()))
-            throw new NullPointerException("FirstName is empty");
+            throw new IllegalArgumentException("FirstName is empty");
         if (entity.getMiddleName() == null)
             throw new NullPointerException("MiddleName is null");
         if ("".equals(entity.getMiddleName()))
-            throw new NullPointerException("MiddleName is empty");
+            throw new IllegalArgumentException("MiddleName is empty");
         if (entity.getBirthday() == null)
             throw new NullPointerException("Birthday is null");
+        if (entity.getExaminations() == null)
+            throw new NullPointerException("Examinations is null");
 
         PatientRecord readEntity = dao.read(entity.getId());
 
-        if(readEntity != null) {
+        if (readEntity != null) {
             dao.update(entity);
             LOGGER.info("PatientRecord(id={}) updated in database", entity.getId());
         } else {
